backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-socketio==5.10.0
python-engineio==4.7.1
ollama==0.1.7
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
websockets==12.0
aiofiles==23.2.1
Pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
asyncio-mqtt==0.16.1
pyyaml==6.0.1
jinja2==3.1.2
prometheus-client==0.19.0
structlog==23.2.0
rich==13.7.0
typer==0.9.0 